Moved helper functions to different file, added test cases for coverage
