+++ src/user/delete.js
@@ -14,6 +14,7 @@ const groups = require('../groups');
 const messaging = require('../messaging');
 const plugins = require('../plugins');
 const batch = require('../batch');
+const makeBulkSet = require('../helpers');
 
 module.exports = function (User) {
 	const deletesInProgress = {};
@@ -207,15 +208,6 @@ module.exports = function (User) {
 			db.getSortedSetRange(`followers:${uid}`, 0, -1),
 			db.getSortedSetRange(`following:${uid}`, 0, -1),
 		]);
-
-		function makeBulkSetElement(uids, index, fieldName, count) {
-			return [`user:${uids[index]}`, { [fieldName]: count || 0 }];
-		}
-
-		function makeBulkSet(counts, uids, fieldName) {
-			console.log('Emily Yu');
-			return counts.map((count, index) => makeBulkSetElement(uids, index, fieldName, count));
-		}
 
 		async function updateCount(uids, name, fieldName) {
 			await batch.processArray(uids, async (uids) => {
